Stock decimal

+++ Database/Tables/Products/Relations/ReturnRequests_Products.sql
@@ -1,5 +1,5 @@
 CREATE TABLE [dbo].[ReturnRequests_Products] (
-    [Quantity]       INT             NOT NULL,
+    [Quantity]       DECIMAL             NOT NULL,
     [UnitPrice]      DECIMAL (18, 2) NOT NULL,
     [Discount]       DECIMAL (5, 2)  NOT NULL,
     [Weight]         INT             NULL,
